+++ src/main/java/de/samply/app/ProjectManagerConst.java
@@ -135,6 +135,7 @@ public class ProjectManagerConst {
     public final static String FETCH_PROJECT_BRIDGEHEAD_RESULTS_FOR_OWN_BRIDGEHEAD_ACTION = "FETCH_PROJECT_BRIDGEHEAD_RESULTS_FOR_OWN_BRIDGEHEAD";
     public final static String FETCH_EMAIL_MESSAGE_AND_SUBJECT_ACTION = "FETCH_EMAIL_MESSAGE_AND_SUBJECT";
     public final static String IS_PROJECT_MANAGER_ADMIN_ACTION = "IS_PROJECT_MANAGER_ADMIN";
+    public final static String FETCH_RESEARCH_ENVIRONMENT_URL_ACTION = "FETCH_RESEARCH_ENVIRONMENT_URL";
 
 
     // REST Services
@@ -237,6 +238,7 @@ public class ProjectManagerConst {
     public final static String FETCH_PROJECT_BRIDGEHEAD_RESULTS_FOR_OWN_BRIDGEHEAD = "/own-project-bridgehead-results";
     public final static String FETCH_EMAIL_MESSAGE_AND_SUBJECT = "/email-message-and-subject";
     public final static String IS_PROJECT_MANAGER_ADMIN = "/is-project-manager-admin";
+    public final static String FETCH_RESEARCH_ENVIRONMENT_URL = "/research-environment-url";
 
 
     // REST Parameters